make userLoop async
